requirements.txt
PyMuPDF==1.25.3
Pillow==11.1.0
tqdm==4.67.1